dom/html/test/test_bug1233598.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1233598
-->
<head>
  <title>Test for Bug 1233598</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1233598">Mozilla Bug 1233598</a>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 1233598 **/

var i;  // must be out here to trigger the leak

function runTest()
{
  i = document.createElement("input");
  i.setAttribute("type", "file");
  i.getFilesAndDirectories();  // returns a promise
  ok(true, "Are we leaking?");
  SimpleTest.finish();
}

SimpleTest.waitForExplicitFinish();
SpecialPowers.pushPrefEnv({"set":[["dom.input.dirpicker", true]]}, runTest);

</script>
</pre>
</body>
</html>
